Updated KeyBinders
Buttons that listen for keyboard input on the options menu now halt hovering graphics when listening for input! There is a 100ms delay to resume graphics (such as the selection box, change in cursor, enabling button) after the player clicks anywhere on the screen so he does not accidentally press a button again.